fixing homework4

+++ service/src/main/java/com/dmdev/dima/entity/Order.java
@@ -10,7 +10,6 @@ import lombok.NoArgsConstructor;
 import lombok.ToString;
 
 import javax.persistence.CascadeType;
-import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
 import javax.persistence.Enumerated;
@@ -18,10 +17,10 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
-import javax.persistence.ManyToMany;
 import javax.persistence.ManyToOne;
 import javax.persistence.OneToMany;
 import javax.persistence.Table;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -39,6 +38,7 @@ public class Order {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+
     @ManyToOne(optional = false)
     @JoinColumn(name = "customer_id")
     private User user;
@@ -46,9 +46,10 @@ public class Order {
     private OrderStatus orderStatus;
     private LocalDateTime orderDate;
     private LocalDateTime payDate;
-    private Float totalSum;
+    private BigDecimal totalSum;
     @Enumerated(EnumType.STRING)
     private Payment payment;
+
     @Builder.Default
     @OneToMany(mappedBy = "order", orphanRemoval = true, cascade = CascadeType.ALL)
     private List<ProductInOrder> productsInOrder = new ArrayList<>();
